app/src/main/java/com/example/tawk/data/db/AppDatabase.kt
package com.example.tawk.data.db

import android.content.Context
import androidx.room.Database
import androidx.room.Room
import androidx.room.RoomDatabase
import com.example.tawk.data.db.entity.RemoteKeys
import com.example.tawk.data.db.entity.User

@Database(
    entities = [User::class, RemoteKeys::class],
    version = 1
)
abstract class AppDatabase : RoomDatabase() {

    abstract fun getUserDao(): UserDao
    abstract fun getRemoteKeysDao(): RemoteKeysDao

    companion object {

        @Volatile
        private var instance: AppDatabase? = null
        private val LOCK = Any()

        operator fun invoke(context: Context) = instance ?: synchronized(LOCK) {
            instance ?: buildDatabase(context).also {
                instance = it
            }
        }

        private fun buildDatabase(context: Context) =
            Room.databaseBuilder(
                context,
                AppDatabase::class.java,
                "Tawkdatabase"
            ).build()


    }
}